+++ tests/update.rs
@@ -1,6 +1,6 @@
 #![cfg(not(feature = "minimal_traits"))]
 
-use aragog::{ServiceError, Update};
+use aragog::{Error, Update};
 use serde::{Deserialize, Serialize};
 
 pub mod common;
@@ -19,9 +19,9 @@ pub struct DishForm {
 }
 
 impl Update<DishForm> for Dish {
-    fn update(&mut self, form: &DishForm) -> Result<(), ServiceError> {
+    fn update(&mut self, form: &DishForm) -> Result<(), Error> {
         if form.price == 0 {
-            return Err(ServiceError::ValidationError(String::from("Wrong price")));
+            return Err(Error::ValidationError(String::from("Wrong price")));
         }
         self.name = form.form_name.clone();
         self.description = form.form_description.clone();